🔧 Handle machines with legacy iptables
On machines using legacy iptables, Docker should use the appropriate
binary (i.e. iptables-legacy instead of iptables-nft), otherwise,
the network can end in a weird state. One symptom is that the
default bridge network works correctly, but custom networks do not
(which prevents us from using e.g. KinD or minikube, since these
will use a custom network).

We're now installing the iptables-legacy package, and we try
to detect if legacy iptables is in use by looking at loaded
modules. Of course, it's still possible that legacy iptables
is compiled into the kernel; but what can you do, eh.

Note: this was made specifically for codespaces, because as of
May 2025, codespaces compute seems to be using legacy iptables.